requirements.txt
python=3.8.10
numpy~=1.20.2
pandas~=1.2.5
matplotlib~=3.3.4
pytorch=1.9.0